move ConsumerSettings to base scaladsl package add create method return CompletableFutures
